Fix welcome handoff lint and comment-prose
Split non-component helpers out of StudioWelcomeView and shrink comments
so the prose baseline and react-refresh rules stay green.
