+++ apps/web/src/router.ts
@@ -105,9 +105,7 @@ export type AppRoute =
   // `/status/:token` is accepted as an alias and resolves here too. Optional `tab`
   // deep-links into a work surface (`/studio/tv-tycoon/build`).
   | { view: 'studio'; game?: string; tab?: StudioTab }
-  // Post–Create Now platform handoff: full-screen wizard chrome before Studio.
-  // Real path so reload / another device resumes the same chapter (not an overlay-only
-  // park). Self/BYOCA connect is a later chapter on its own path.
+  // Platform create handoff before Studio chrome.
   | { view: 'studioWelcome'; game: string }
   // Privacy policy and terms. Reachable without a session — someone deciding whether
   // to sign in has to be able to read what signing in would mean first.
@@ -302,10 +300,9 @@ export function parsePathRoute(pathname: string, hash = ''): AppRoute {
     return { view: 'proposals' };
   }
 
-  // `/studio/:game` or `/studio/:game/:tab` (or `:chapter` for post-create handoff).
-  // A third segment that is not a known tab/chapter is a 404 rather than a silent
-  // fallback to the game-only view: it keeps the client in step with the API's shell
-  // allowlist, which serves those paths a real 404.
+  // `/studio/:game` or `/studio/:game/:tab`. A third segment that is not a known
+  // tab is a 404 rather than a silent fallback to the game-only view: it keeps the
+  // client in step with the API's shell allowlist, which serves those paths a real 404.
   const studioMatch = normalizedPath.match(/^\/studio\/([^/]+)(?:\/([^/]+))?$/);
   if (studioMatch?.[1]) {
     const game = decodeSegment(studioMatch[1]);
@@ -316,8 +313,7 @@ export function parsePathRoute(pathname: string, hash = ''): AppRoute {
     if (!tabSegment) {
       return { view: 'studio', game };
     }
-    // Platform create handoff — not a Studio work surface; its own view so App can
-    // mount wizard chrome without the shelf.
+    // Platform create handoff — wizard chrome, not a Studio tab.
     if (tabSegment === 'welcome') {
       return { view: 'studioWelcome', game };
     }
@@ -464,7 +460,7 @@ export function studioPath(game?: string, tab?: StudioTab): string {
   return tab ? `${base}/${tab}` : base;
 }
 
-/** Full-screen platform handoff after Create Now — before Studio chrome. */
+/** Full-screen platform handoff after Create Now. */
 export function studioWelcomePath(game: string): string {
   return `/studio/${encodeURIComponent(game)}/welcome`;
 }
@@ -502,8 +498,7 @@ export function navUpTarget(route: AppRoute): NavUpTarget | null {
       }
       return { path: '/', labelKey: 'upHome' };
     case 'studioWelcome':
-      // Still in the create flow — Up leaves the handoff for the shelf, same as a
-      // selected-game Studio URL, rather than dumping home mid-build.
+      // Up leaves handoff for the shelf.
       return { path: studioPath(), labelKey: 'upStudio' };
     case 'game':
       // The page is nested under the creator profile the way a repo nests under its
