+++ eslint-rules/comment-prose-baseline.json
@@ -54,7 +54,7 @@
     "apps/api/src/assemble.test.ts": 220,
     "apps/api/src/assemble.ts": 272,
     "apps/api/src/auth.test.ts": 144,
-    "apps/api/src/auth.ts": 1998,
+    "apps/api/src/auth.ts": 1931,
     "apps/api/src/autonomy.test.ts": 182,
     "apps/api/src/autonomy.ts": 474,
     "apps/api/src/bearer.ts": 35,
@@ -176,7 +176,7 @@
     "apps/api/src/mcp-session-nudges.test.ts": 17,
     "apps/api/src/mcp-session-nudges.ts": 494,
     "apps/api/src/mcp-ui-flag.test.ts": 224,
-    "apps/api/src/mcp-ui.test.ts": 2081,
+    "apps/api/src/mcp-ui.test.ts": 2059,
     "apps/api/src/mcp-ui.ts": 1803,
     "apps/api/src/moderation-metrics.test.ts": 583,
     "apps/api/src/moderation-metrics.ts": 489,
@@ -321,9 +321,9 @@
     "apps/api/vitest.setup.ts": 122,
     "apps/e2e/src/anonymous-and-mobile.test.ts": 136,
     "apps/e2e/src/auth.test.ts": 285,
-    "apps/e2e/src/browser.ts": 1364,
+    "apps/e2e/src/browser.ts": 1290,
     "apps/e2e/src/config.ts": 180,
-    "apps/e2e/src/games-playable.test.ts": 726,
+    "apps/e2e/src/games-playable.test.ts": 712,
     "apps/e2e/src/gate-prerequisites.test.ts": 164,
     "apps/e2e/src/globalSetup.ts": 153,
     "apps/e2e/src/resilience.test.ts": 356,
@@ -344,12 +344,12 @@
     "apps/web/src/AdminJobsPanel.tsx": 587,
     "apps/web/src/api.ts": 46,
     "apps/web/src/App.anonymousPlay.test.ts": 40,
-    "apps/web/src/App.catalog.test.ts": 131,
+    "apps/web/src/App.catalog.test.ts": 110,
     "apps/web/src/App.createGameFocus.test.ts": 33,
     "apps/web/src/App.gamePageTheater.test.ts": 86,
-    "apps/web/src/App.overlayExit.test.ts": 62,
+    "apps/web/src/App.overlayExit.test.ts": 52,
     "apps/web/src/App.qa.test.ts": 340,
-    "apps/web/src/App.tsx": 2567,
+    "apps/web/src/App.tsx": 2500,
     "apps/web/src/AppleSignInButton.test.tsx": 110,
     "apps/web/src/AppleSignInButton.tsx": 268,
     "apps/web/src/AppUpdateBanner.test.tsx": 36,
@@ -395,12 +395,12 @@
     "apps/web/src/EditorPanel.tsx": 600,
     "apps/web/src/EditProfileModal.tsx": 56,
     "apps/web/src/GameDetailPage.test.tsx": 15,
-    "apps/web/src/GameDetailPage.tsx": 122,
+    "apps/web/src/GameDetailPage.tsx": 36,
     "apps/web/src/GameFrame.sandbox.test.ts": 34,
     "apps/web/src/GameFrame.tsx": 448,
     "apps/web/src/GameHealthView.test.tsx": 302,
     "apps/web/src/GameHealthView.tsx": 494,
-    "apps/web/src/GamePage.tsx": 114,
+    "apps/web/src/GamePage.tsx": 87,
     "apps/web/src/gamePageApi.ts": 59,
     "apps/web/src/gamePlayer.bridge.test.ts": 334,
     "apps/web/src/gamePlayer.hiddenFrame.test.tsx": 97,
@@ -502,7 +502,7 @@
     "apps/web/src/ReportGameButton.test.tsx": 137,
     "apps/web/src/ReportGameButton.tsx": 180,
     "apps/web/src/router.test.ts": 366,
-    "apps/web/src/router.ts": 2204,
+    "apps/web/src/router.ts": 2172,
     "apps/web/src/ScorecardPanel.test.tsx": 185,
     "apps/web/src/ScorecardPanel.tsx": 388,
     "apps/web/src/selfBuildCopy.test.ts": 102,
@@ -549,12 +549,12 @@
     "apps/web/src/telemetry.test.ts": 201,
     "apps/web/src/telemetry.ts": 981,
     "apps/web/src/TelemetryCharts.test.tsx": 26,
-    "apps/web/src/TelemetryCharts.tsx": 296,
+    "apps/web/src/TelemetryCharts.tsx": 289,
     "apps/web/src/TelemetryOverview.test.tsx": 29,
     "apps/web/src/TelemetryOverview.tsx": 131,
     "apps/web/src/telemetryTrends.ts": 81,
     "apps/web/src/TelemetryTrendsPanel.tsx": 49,
-    "apps/web/src/UnpublishedPlayView.tsx": 155,
+    "apps/web/src/UnpublishedPlayView.tsx": 126,
     "apps/web/src/useDeviceTilt.test.tsx": 328,
     "apps/web/src/useDeviceTilt.ts": 809,
     "apps/web/src/useDocumentTitle.ts": 34,
